chatbot/requirement.txt
fastapi
uvicorn
groq
motor
python-dotenv
pydantic